Backend/requirements.txt
django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
